+++ packages/egenie-utils/src/programme/filterItems/filterDate.tsx
@@ -1,13 +1,13 @@
 import { DatePicker, Select, Typography } from 'antd';
 import classNames from 'classnames';
 import _ from 'lodash';
-import { action, set, observable, toJS } from 'mobx';
+import { action, observable, extendObservable, toJS } from 'mobx';
 import { observer } from 'mobx-react';
 import moment from 'moment';
 import React from 'react';
 import { ENUM_FILTER_ITEM_TYPE, FilterBase } from './common';
 
-enum FormatType {
+export enum FormatType {
   defaultFormat = 'YYYY-MM-DD HH:mm:ss',
   day = 'YYYY-MM-DD',
 }
@@ -39,11 +39,19 @@ function formatTime(startTime: moment.Moment, endTime: moment.Moment, format: st
   }
 }
 
+/**
+ * 时间范围
+ */
 export class FilterDate extends FilterBase {
   constructor(options: Partial<FilterDate>) {
-    super();
-    set(this, {
-      ...options,
+    super(options);
+    const {
+      data,
+      ...rest
+    } = options;
+
+    extendObservable(this, {
+      ...rest,
       showCollapse: false,
     });
     this.snapshot = {
@@ -54,17 +62,12 @@ export class FilterDate extends FilterBase {
   }
 
   /**
-   * 是否显示更多时间字段项目
-   */
-  @observable public showSelect = true;
-
-  /**
-   * 类型标志
+   * 类型标志。date为左右查询方案类型(带下拉选择,左右查询方案)。dateRange。DatePicker.Range
    */
   @observable public type: 'date' | 'dateRange' = ENUM_FILTER_ITEM_TYPE.date;
 
   public toProgramme(): string | null {
-    if (this.showSelect) {
+    if (this.type === ENUM_FILTER_ITEM_TYPE.date) {
       if (this.selectValue) {
         return `${this.selectValue},${formatTime(this.startTime, this.endTime, this.format)}`;
       } else {
@@ -77,7 +80,7 @@ export class FilterDate extends FilterBase {
 
   public toParams(this: FilterDate): {[key: string]: string; } {
     const timeString = formatTime(this.startTime, this.endTime, this.format);
-    if (this.showSelect) {
+    if (this.type === ENUM_FILTER_ITEM_TYPE.date) {
       if (this.selectValue) {
         if (timeString) {
           return {
@@ -103,7 +106,7 @@ export class FilterDate extends FilterBase {
   public formatValue(this: FilterDate, value?: string): void {
     const keyAndValues = _.toString(value)
       .split(',');
-    if (this.showSelect) {
+    if (this.type === ENUM_FILTER_ITEM_TYPE.date) {
       this.selectValue = keyAndValues[0] ? keyAndValues[0] : undefined;
       this.startTime = keyAndValues[1] ? moment(keyAndValues[1]) : null;
       this.endTime = keyAndValues[2] ? moment(keyAndValues[2]) : null;
@@ -127,9 +130,22 @@ export class FilterDate extends FilterBase {
   };
 
   /**
+   * 允许起始项部分为空。type为dateRange生效
+   */
+  @observable public allowEmpty: [boolean, boolean] = [
+    true,
+    true,
+  ];
+
+  /**
    * 日期类型选中值
    */
   @observable public selectValue: string | undefined = undefined;
+
+  /**
+   * 日期改变回掉
+   */
+  public handleChangeCallback: (date: [moment.Moment, moment.Moment]) => void;
 
   /**
    * @internal
@@ -146,19 +162,40 @@ export class FilterDate extends FilterBase {
   /**
    * 开始时间
    */
-  @observable public startTime: moment.Moment | null = null;
+  @observable.ref public startTime: moment.Moment | null = null;
 
   /**
    * @internal
    */
   @action public handleStartChange = (startTime: moment.Moment | null) => {
     this.startTime = startTime;
+
+    if (typeof this.handleChangeCallback === 'function') {
+      this.handleChangeCallback([
+        this.startTime,
+        this.endTime,
+      ]);
+    }
+  };
+
+  /**
+   * @internal
+   */
+  @action public handleRangeChange = (dates: [moment.Moment, moment.Moment]) => {
+    this.startTime = dates?.[0];
+    this.endTime = dates?.[1];
+    if (typeof this.handleChangeCallback === 'function') {
+      this.handleChangeCallback([
+        this.startTime,
+        this.endTime,
+      ]);
+    }
   };
 
   /**
    * 结束时间
    */
-  @observable public endTime: moment.Moment | null = null;
+  @observable.ref public endTime: moment.Moment | null = null;
 
   /**
    * 禁止状态
@@ -173,6 +210,13 @@ export class FilterDate extends FilterBase {
    */
   @action public handleEndChange = (endTime: moment.Moment | null) => {
     this.endTime = endTime;
+
+    if (typeof this.handleChangeCallback === 'function') {
+      this.handleChangeCallback([
+        this.startTime,
+        this.endTime,
+      ]);
+    }
   };
 
   /**
@@ -188,7 +232,21 @@ export class FilterDate extends FilterBase {
  * @internal
  */
 @observer
-export class FilterDateWrapperComponent extends React.Component<{ store: FilterDate; }> {
+export class FilterDateComponent extends React.Component<{ store: FilterDate; }> {
+  render() {
+    if (this.props.store.type === ENUM_FILTER_ITEM_TYPE.date) {
+      return <FilterDateNormal store={this.props.store}/>;
+    } else {
+      return <FilterDateRange store={this.props.store}/>;
+    }
+  }
+}
+
+/**
+ * @internal
+ */
+@observer
+class FilterDateNormal extends React.Component<{ store: FilterDate; }> {
   render() {
     const {
       startTime,
@@ -203,10 +261,9 @@ export class FilterDateWrapperComponent extends React.Component<{ store: FilterD
       label,
       className,
       style,
-      showSelect,
       disabled,
     } = this.props.store;
-    const newClassName = classNames(showSelect ? 'filterDateSelect' : 'filterDateNormal', className);
+    const newClassName = classNames('filterDateSelect', className);
     return (
       <div
         className={newClassName}
@@ -221,17 +278,13 @@ export class FilterDateWrapperComponent extends React.Component<{ store: FilterD
               {label}
             </Typography.Title>
           </section>
-          {
-            showSelect ? (
-              <Select
-                bordered={false}
-                onChange={handleSelectChange}
-                options={data}
-                placeholder="请选择"
-                value={selectValue}
-              />
-            ) : null
-          }
+          <Select
+            bordered={false}
+            onChange={handleSelectChange}
+            options={data}
+            placeholder="请选择"
+            value={selectValue}
+          />
         </header>
         <section>
           <DatePicker
@@ -248,13 +301,9 @@ export class FilterDateWrapperComponent extends React.Component<{ store: FilterD
             suffixIcon={null}
             value={startTime}
           />
-          {
-            showSelect ? (
-              <div>
-                至
-              </div>
-            ) : null
-          }
+          <div>
+            至
+          </div>
           <DatePicker
             bordered={false}
             disabled={disabled[1]}
@@ -265,10 +314,68 @@ export class FilterDateWrapperComponent extends React.Component<{ store: FilterD
               hideDisabledOptions: true,
               defaultValue: moment('23:59:59', 'HH:mm:ss'),
             } : false}
-            showToday={format !== FormatType.defaultFormat}
             value={endTime}
           />
-          <span/>
+        </section>
+      </div>
+    );
+  }
+}
+
+/**
+ * @internal
+ */
+@observer
+class FilterDateRange extends React.Component<{ store: FilterDate; }> {
+  render() {
+    const {
+      allowEmpty,
+      startTime,
+      endTime,
+      placeholder,
+      format,
+      label,
+      className,
+      style,
+      disabled,
+      handleRangeChange,
+    } = this.props.store;
+    const newClassName = classNames('filterDateNormal', className);
+    return (
+      <div
+        className={newClassName}
+        style={toJS(style)}
+      >
+        <header>
+          <section className="filterLabel">
+            <Typography.Title
+              ellipsis={{ rows: 1 }}
+              title={label}
+            >
+              {label}
+            </Typography.Title>
+          </section>
+        </header>
+        <section>
+          <DatePicker.RangePicker
+            allowEmpty={allowEmpty}
+            bordered={false}
+            disabled={disabled}
+            format={format}
+            onChange={handleRangeChange}
+            placeholder={placeholder}
+            showTime={format === FormatType.defaultFormat ? {
+              hideDisabledOptions: true,
+              defaultValue: [
+                moment('00:00:00', 'HH:mm:ss'),
+                moment('23:59:59', 'HH:mm:ss'),
+              ],
+            } : false}
+            value={[
+              startTime,
+              endTime,
+            ]}
+          />
         </section>
       </div>
     );