+++ packages/egenie-utils/src/programme/filterItems/filterInputOrSelect.tsx
@@ -1,6 +1,6 @@
 import { Input, Select } from 'antd';
 import _ from 'lodash';
-import { action, set, observable, toJS } from 'mobx';
+import { action, extendObservable, observable, toJS } from 'mobx';
 import { observer } from 'mobx-react';
 import React from 'react';
 import { ENUM_FILTER_ITEM_TYPE, FilterBase } from './common';
@@ -8,9 +8,12 @@ import { ENUM_FILTER_ITEM_TYPE, FilterBase } from './common';
 export class FilterInputOrSelect extends FilterBase {
   constructor(options: Partial<FilterInputOrSelect>) {
     super(options);
-    set(this, {
-      toParams: this.toParams,
-      ...options,
+    const {
+      data,
+      ...rest
+    } = options;
+    extendObservable(this, {
+      ...rest,
       showCollapse: false,
     });
     this.formatValue(this.value);