+++ packages/egenie-utils/src/programme/filterItems/filterSelect.tsx
@@ -1,5 +1,5 @@
 import { Button, Divider, Select, Typography } from 'antd';
-import { action, computed, set, intercept, observable, toJS } from 'mobx';
+import { action, computed, extendObservable, intercept, observable, toJS } from 'mobx';
 import { observer } from 'mobx-react';
 import React from 'react';
 import { ENUM_FILTER_ITEM_TYPE, FilterBase, ValueAndLabelData } from './common';
@@ -23,9 +23,12 @@ function formatValue(oldValue, mode) {
 export class FilterSelect extends FilterBase {
   constructor(options: Partial<FilterSelect>) {
     super(options);
-    set(this, {
-      toParams: this.toParams,
-      ...options,
+    const {
+      data,
+      ...rest
+    } = options;
+    extendObservable(this, {
+      ...rest,
       showCollapse: false,
     });
     if (!this.placeholder) {